requirements.txt
requests~=2.28.1
selenium~=4.4.0
python-dotenv~=0.20.0
